refactor: Extract change password link
